Link native build nodes into eclipse.platform and equinox JIPP
The 'remoteFS' of the agents is not individualized since for each job
Jenkins creates an individual workspace path based on the job's name.
And since the jobs in platform.releng.aggregator, platform and equinox
already have individual names there is no overlap.
Leveraging that helps to keep the path's short, which is still necessary
on Windows.

Required for
https://gitlab.eclipse.org/eclipsefdn/helpdesk/-/issues/4231