Link native build nodes into eclipse.platform and equinox JIPP
Besides linking the 'jenkins/configuration.yml' to the eclipse.platform
and equinox JIPP, this introduce the placeholder '&&INSTANCE_NAME&&'
which is replaced with the project's full name while building images.
In this case this is used to ensure different workspace locations for
the different JIPPs that use the native nodes.

Required for
https://gitlab.eclipse.org/eclipsefdn/helpdesk/-/issues/4231